Implement xmlattr access and type enforcement
This adds an :access field to xmlattrs, so an attr can be readonly -
that is, it won't have an attr_writer method. In addition, the newly
added method does type enforcement and updates the resource state. So
finally, the status method for entries is actually meaningful!

This fixes (gprov #1) and (gprov #3).